LPFG-569: Renamed MediaFactory to MediaEntityFactory
