+++ src/main/java/uk/gov/cslearning/catalogue/domain/media/MediaEntityFactory.java
@@ -1,9 +1,7 @@
-package uk.gov.cslearning.catalogue.dto.factory;
+package uk.gov.cslearning.catalogue.domain.media;
 
 import com.google.common.collect.ImmutableMap;
 import org.springframework.stereotype.Component;
-import uk.gov.cslearning.catalogue.domain.media.Document;
-import uk.gov.cslearning.catalogue.domain.media.MediaEntity;
 import uk.gov.cslearning.catalogue.dto.FileUpload;
 import uk.gov.cslearning.catalogue.exception.UnknownFileTypeException;
 
@@ -12,7 +10,7 @@ import java.util.Map;
 import java.util.function.Function;
 
 @Component
-public class MediaFactory {
+public class MediaEntityFactory {
 
     private Map<String, Function<FileUpload, MediaEntity>> createMethods = ImmutableMap.of(
             "doc", new CreateDocumentFunction()
