+++ src/main/java/uk/gov/cslearning/catalogue/service/AzureMediaManagementService.java
@@ -7,8 +7,8 @@ import com.microsoft.azure.storage.blob.CloudBlockBlob;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import uk.gov.cslearning.catalogue.domain.media.Media;
+import uk.gov.cslearning.catalogue.domain.media.MediaEntityFactory;
 import uk.gov.cslearning.catalogue.dto.FileUpload;
-import uk.gov.cslearning.catalogue.dto.factory.MediaFactory;
 import uk.gov.cslearning.catalogue.exception.FileUploadException;
 import uk.gov.cslearning.catalogue.repository.MediaRepository;
 
@@ -19,11 +19,11 @@ import java.net.URISyntaxException;
 public class AzureMediaManagementService implements MediaManagementService {
 
     private final CloudBlobClient azureClient;
-    private final MediaFactory mediaFactory;
+    private final MediaEntityFactory mediaFactory;
     private final MediaRepository mediaRepository;
     private final String storageContainerName;
 
-    public AzureMediaManagementService(CloudBlobClient azureClient, MediaFactory mediaFactory, MediaRepository mediaRepository, @Value("${azure.storage.container}") String storageContainerName) {
+    public AzureMediaManagementService(CloudBlobClient azureClient, MediaEntityFactory mediaFactory, MediaRepository mediaRepository, @Value("${azure.storage.container}") String storageContainerName) {
         this.azureClient = azureClient;
         this.mediaFactory = mediaFactory;
         this.mediaRepository = mediaRepository;
